Create separate SubwayNavStep component
