+++ packages/dashboard/src/components/SubwayNav/SubwayNav.test.tsx
@@ -3,7 +3,7 @@ import * as renderer from 'react-test-renderer';
 import { mount } from 'enzyme';
 
 import { SubwayNav } from './SubwayNav';
-import { ISubwayNavStep } from '@uifabric/dashboard/lib/components/SubwayNav/SubwayNav.types';
+import { ISubwayNavStep } from '@uifabric/dashboard/lib/components/SubwayNav/SubwayNavStep.types';
 
 /**
  * generate Random id
