+++ packages/dashboard/src/components/SubwayNav/examples/SubwayNav.Basic.DifferentStates.Example.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { generateRandomId } from '@uifabric/dashboard/lib/components/SubwayNav/examples/SubwayNav.Util';
 import { SubwayNav } from '@uifabric/dashboard/lib/components/SubwayNav/SubwayNav';
-import { ISubwayNavStep, SubwayNavStepState } from '@uifabric/dashboard/lib/components/SubwayNav/SubwayNav.types';
+import { ISubwayNavStep, SubwayNavStepState } from '@uifabric/dashboard/lib/components/SubwayNav/SubwayNavStep.types';
 
 export class SubwayNavBasicDifferentStatesExample extends React.Component<any, any> {
   public render(): JSX.Element {
